requirements.txt
asgiref==3.5.2
Django==4.1.3
sqlparse==0.4.3
psycopg2
gunicorn
django-environ
django-cors-headers
djangorestframework
Pillow
django-rest-swagger
